perf report generation for device perf

+++ tools/tracy/__main__.py
@@ -413,8 +413,15 @@ def main():
             logger.info(f"Test process started")
 
             def signal_handler(sig, frame):
-                os.killpg(os.getpgid(testProcess.pid), signal.SIGTERM)
-                captureProcess.terminate()
+                try:
+                    if testProcess.poll() is None:
+                        os.killpg(os.getpgid(testProcess.pid), signal.SIGTERM)
+                except ProcessLookupError:
+                    pass
+                try:
+                    captureProcess.terminate()
+                except ProcessLookupError:
+                    pass
                 captureProcess.communicate()
                 sys.exit(3)
